Add release helper module with CLI and core functionalities
- Implemented the release helper module with various utilities for managing releases in a monorepo.
- Added command line interface (CLI) for interacting with the release helper, including commands for listing apps, building images, releasing, and validating.
- Introduced change detection utilities to identify modified applications since a specified Git tag.
- Developed core functionalities for finding workspace root and running Bazel commands.
- Created Git operations for tagging and pushing images to a container registry.
- Implemented image building and tagging utilities for managing container images.
- Added metadata utilities for retrieving app metadata and listing all apps in the monorepo.
- Developed release planning and execution utilities to facilitate the release process.
- Included summary generation utilities for creating release summaries for GitHub Actions.
- Implemented validation utilities to ensure semantic versioning and app existence checks.
