+++ tools/release_helper/__init__.py
@@ -1,0 +1,1 @@
+# Release helper module
